Add variant of the gbuffer and gbuffer material pass that take a depth buffer from a pre-z pass.

+++ Source/Core/RenderDevice.cpp
@@ -1062,7 +1062,7 @@ void RenderDevice::execute(RenderGraph& graph)
             passBegin.setRenderArea(renderArea);
             if(!clearValues.empty())
             {
-                passBegin.setClearValueCount(clearValues.size());
+                passBegin.setClearValueCount(static_cast<uint32_t>(clearValues.size()));
                 passBegin.setPClearValues(clearValues.data());
             }
